Bug 1638533 - Add (in promise) suffix for uncaught rejected promise. r=davidwalsh.
Now that nsIScriptError have a `isPromiseRejection` flag, we can consume it
in the console to indicate to the user that we're dealing with a promise rejection.
The exception message is then: "Uncaught (in promise) XXX".
A few tests needed to be updated, and we take this as an opportunity to
ensure we have a stacktrace for uncaught promise rejection.

Differential Revision: https://phabricator.services.mozilla.com/D75780

UltraBlame original commit: bc1d9c3a316731a48dccbe06b6b96942b13ef9a5