Bug 1629636 - Make Windows install not register "ftp" protocol handler for NIGHTLY_BUILD. r=mhowell
For builds with ftp disabled (see below), this commit:

1) stops registering the ftp protocol handler at install time;
2) actively unregisters the ftp protocol handler at postupdate time;
3) stops unregistering the ftp protocol handler at uninstall time.

The rationale for 3) is that by the time a `helper.exe` with this
change is in place, the postupdate step has already run and
unregistered the ftp protocol handler.  This could, of course, fail,
and a fallback would be nice.  However having a guarded block, just
like everywhere else, will make it much more likely that the complete
removal of the ftp protocol will also cull the uninstall code.  I
prefer making the latter cleanup more likely to be complete.

The bool pref that disables ftp functionality is
"network.ftp.enabled", and at this time that defaults to
!NIGHTLY_BUILD.  In the {un}install process, there's no way to inspect
that pref dynamically, so we use !NIGHTLY_BUILD as well.

This opens a race window for developers to change the pref default
without changing the {un}install conditional at the same time.  It
would be possible to close that window by introducing a new configure
subst but given the imminent removal of the ftp protocol entirely it
doesn't seem necessary.

Differential Revision: https://phabricator.services.mozilla.com/D74503

UltraBlame original commit: 741e5111fa4019d8d2ba3a80c72521c6a8590f2a