Bug 1638574 - Allow to use widl when cross-compiling clang-cl builds. r=glandium
This avoids having to have clang-cl.exe and ml64.exe in the path.

This requires a WIDL patch[1] and one other tweak, but it's working
other than that.

While at it, allow to pass extra flags to midl / widl. This is useful to
test with a local widl for example, to pass the right include paths.

[1]: https://www.winehq.org/pipermail/wine-devel/2020-May/166368.html

Differential Revision: https://phabricator.services.mozilla.com/D75680

UltraBlame original commit: 7054e490f0ce9a4fe79e7047dbde56c014eea223