Bug 1628073 - Rewrite view_gecko_profile.py to open a URL; r=perftest-reviewers,sparky
The current implementation opens the built Firefox in the objdir. This is not
optimal as the built Firefox is not really great for viewing files in. The build
could be broken. With this patch, the profiles will instead be opened in the users
default browser.

Differential Revision: https://phabricator.services.mozilla.com/D70089

UltraBlame original commit: 448ea7e8873c622a1fa4f70f93503f4cedb84c39