Bug 1639825 - Some minor code reorganization. r=tnikkel
This makes it more obvious that we are using the same scrollDelta for adjusting
the displayport as we are for populating the callback transform. Instead of
computing the same value inside a helper, we just provide the helper with the
already-computed delta that is used for the callback transform too.

Differential Revision: https://phabricator.services.mozilla.com/D76297

UltraBlame original commit: cd4b7928b25c5ffe59863139e5e440fbc31f084a