Bug 1639988 - Disable the blocking from bug 1525084. r=aosmond
This blocking is preventing a lot of users from getting WebRender and it
doesn't seem like the problem is actually reproducible in a lot of
situations. By disabling the blocking I'd like to see if we can get more
reports so that we can properly tighten the list.

Differential Revision: https://phabricator.services.mozilla.com/D76390

UltraBlame original commit: d5255379819c0c4577eb4cbe46c8bb141a3d21d2