Bug 1636192 - Clean Content process Target actor on process shutdown. r=nchevobbe
Before this patch, we were not trying to cleanup anything when the process
was known to be in process of destruction. We assumed that everything would
be cleaned up anyway.
But mochitest is actually asserting that nothing is leaked on process shutdown,
and we are easily leaking many thing when not destroying the target actor!

Differential Revision: https://phabricator.services.mozilla.com/D76156

UltraBlame original commit: d45a9f533b347b9370c60ab6c8ec1329eb8d105c