Bug 1640158 - Remove useless principal-passing in the CSS loader. r=ckerschb
The only caller that passes that principal is the only caller that calls
into LoadSheet with a loader that has a document anyway.

And the principal we're passing is Document::NodePrincipal(), which is
what we'd end up using anyway, so the new code is exactly equivalent, as
far as I can tell.

Differential Revision: https://phabricator.services.mozilla.com/D76469

UltraBlame original commit: 9b3266bb0f915f1a033ef91cc50433ded0764efe