Bug 1638664 - Run the windows searchfox build as cross-compilation on Linux. r=asuth
This updates the task definition for the win64 searchfox job to closely match
the win64/debug task definition in taskcluster/ci/build/windows.yml. So, instead
of running the build with the mozsearch-plugin on a windows worker, it runs
on a Linux worker and does a cross-compile of windows code. The end result in
terms of searchfox artifacts is equivalent, except for absolute filename paths.
I verified that with mozsearch/mozsearch#299 and mozsearch/mozsearch-mozilla#87
in place, searchfox correctly indexes windows-only C++ and Rust code.

Differential Revision: https://phabricator.services.mozilla.com/D76721

UltraBlame original commit: 56729c1c06ba37ea8614025e3ccc184e222f2aee