Bug 1640381. Make ViewportUtils::IsZoomedContentRoot handle print preview documents. r=botond
In print preview documents the child of the root scroll frame is a page sequence frame.

I did a basic test of pinch zooming and panning (in a debug build and in an opt build) around on a print preview document and it seemed to work fine.

Differential Revision: https://phabricator.services.mozilla.com/D76591

UltraBlame original commit: f5409e1328456446d77187dbfcf7cfe824985ab8