Bug 1640140 - Fix rendering of very large regexp. r=Honza.
If the displayString of the RegExp is very long (>10,000 chars),
it is returned as a longString grip, and it was making the regexp
reps to throw.
This patch fixes this issue and adds a test to ensure we deal with
these kind of objects now.

Depends on D76673

Differential Revision: https://phabricator.services.mozilla.com/D76674

UltraBlame original commit: 9298de7e11dd1a8b7cf408d1f3fba6f0edb6c55e