Bug 1599858 - Defer JSFunction::setTypeForScriptedFunction until after BCE. r=mgaudet
We defer the call until after BytecodeEmitter finishes running but before the
ScriptStencils are instantiated. We use the FunctionBox::wasEmitted flag to
know the function is now referenced by bytecode in the enclosing script and
that it is time to apply the complete ObjectGroup information.

Differential Revision: https://phabricator.services.mozilla.com/D76548

UltraBlame original commit: c1496d6074bc3a6bbf11797915d72269e7e8ae79